visualizeUtils.js
